requirements.txt
Flask
PyPDF2
Pillow
opencv-python
pdf2image
gunicorn
